Raise if additional workers process locked documents
These workers do not inherit from PublishingApiWorker, so we add individual
checks for locked documents to these workers and raise an error if the document
is locked. They both inherit from WorkerBase, however this is inherited by many
other unrelated workers, so it doesn't feel appropriate to add a shared locked
document check to that class.