Add DocumentCollectionNonWhitehallLink model
This model is intended to store the data of GOV.UK content that is not
stored in Whitehalls local database. It has been created as a means to
associate that content with a document collection, which will enable
them to use content from other publishing applications.

The fields stored on the model are, content_id, base_path, title and
publishing_app. It is expected that base_path, title and publishing_app
will be shown to the user in the Whitehall user interface, whereas the
content_id will be sent to the Publishing API. The base_path, title and
publishing_app values may slip out of date.

The expectation is that a URL (or base_path) to GOV.UK can be provided
by a user which is then looked up in the Publishing API and other data
saved.

This is something of a rapid band-aid approach to solve this problem as
ideally the whole document collection process should be powered by
Publishing API data long term.