Only apply 'words to avoid' feature to edition[body] textareas
Currently Whitehall hijacks any use of `textarea` and applies this
'words to avoid' checker to it, which wraps the textarea in a div
and overlays a highlighter to highlight any forbidden words.

This is a reasonable/useful feature for the main edition body, but
less applicable to auxiliary fields such as image caption, and
its JavaScript-driven nature interferes with any other behaviour
we might have for that textarea (e.g. show/hide based on radio
button status).

It seems reasonable to use a more specific selector to initialise
the library with.