Replace govuk_frontend_toolkit media queries
This has been done as part of the process to removing the dependency on
gouvk_frontend_toolkit. It migrates media queries to use the helper
function that govuk-frontend provides.

These media queries are now based on em values rather than pixels. A
consideration in the admin side of Whitehall is that it uses bootstrap 2
as the styling basis which has a body font-size of 14px. We need to
specify this as a variable in admin/base.scss so that the em
calculations consider an em to be 14px in size (rather than the more
common 16px).