Remove IE specific code from JS files
This removes specific JavaScript conditionals that are based on the
version of IE a user is browsing with. These were used to target < IE 10
which is rather legacy, as per the service manual we only actively test
and support IE 11 and up.

These browser checks are somewhat intertwined with parts of
govuk_frontend_toolkit which provides helper functions for SCSS files
and IE. The govuk_frontend_toolkit integration is in the process of
being removed and the equivalent SCSS helpers will no longer exist.