Remove CSS rules only for legacy IE versions
We are removing govuk_frontend_toolkit from this project as it
deprecated and no-longer under development. One of the features of this
project was helper functions to target particular IE versions. When we
remove govuk_frontend_toolkit we will no longer have these available.

As these target very old versions of IE (< 10) it seems unnecessary to
keep these as they are no longer browsers we actively support (as per
the service manual [1] IE 11 is the oldest IE recommended for testing).

Removing these frees us to remove the individual stylesheets for each IE
version which should improve asset generation speed significantly.

[1]: https://www.gov.uk/service-manual/technology/designing-for-different-browsers-and-devices