Add `#english?` to check if primary_locale is "en"
We currently only have `#non_english?`, which is sometimes used to check
if an edition's `primary_locale` is "en", negating it: `!non_english?`.
This method will be used in a later commit to change those instances
from `!non_english?` to `english?`, which is much easier to visually and
mentally parse.