Use id as slug if locale uses a non latin script
We recently allowed DocumentCollection & Consultation to change locale:
- https://github.com/alphagov/whitehall/pull/5918
- https://github.com/alphagov/whitehall/pull/5931

In both cases, we ensured that slugs are always created using the title,
regardless of locale:
- https://github.com/alphagov/whitehall/pull/5918/commits/f6d7550dae7630101cadf440c51f81700c52076f
- https://github.com/alphagov/whitehall/pull/5931/commits/a0d22f2360bff1d664d651641df3de66f8443791

This commit changes that behaviour to only allow locales that use a
latin script alphabet to create the slug from the title. For languages
that use a non latin script alphabet, `string_for_slug` will return nil
so the document id will be used as slug.