Fix document_collection and consultation tests
For document_collection we were testing that the collection's document's
slug was equal to the collection's title. Similarly, for consultation we
were testing that the consultation's document's slug was equal to the
consultation's title.

This is incorrect and the tests were only passing because of how we were
creating the factory objects. Their titles were created with only lower
case and hyphenated strings. After creating more realistic titles in the
previous commit and removing the title overrides in the tests, they
started failing.

The document's slugs of both document_collection and consultation depend
on the title of the document_collection and consultation respectively,
but they are not equal. We use the friendly_id gem to create the slugs
by editing the titles. We also modify the standard gem's behaviour in
lib/friendly_id/custom_normalise.rb, where we can see how the titles
become slugs.

I considered using `#parameterize` in the tests to avoid having to call
`#normalize_friendly_id` but although the tests pass in many cases, they
crucially fail when the titles contain some special characters. For
example, with title `Edition's title (#1): colon, comma and £ sign` we
get:

```
> document.normalize_friendly_id(collection.title)
=> "editions-title-1-colon-comma-and-sign"

> collection.title.parameterize
=> "edition-s-title-1-colon-comma-and-sign"
```