Render placeholder for lead image if assets are missing
We can tap into the placeholder behaviour; we are sending a placeholder when composing the payload, so we should render that too.

We are also sending the placeholder when assets are missing for the custom lead image, but we're currently not rendering it fully, only include it in the select. Perhaps it should be excluded from the select if assets are missing/still processing.