Simplify `Ga4FormTracker`
Add support for using `ga4-form-section` to determine section of changed
input. This means that it is easier to determine the section of elements
nested within `fieldset` elements or elements that represent a single
field spread across multiple elements.

Remove `getJSON` and use the `findTrackingAttributes` from
`GOVUK.analyticsGa4.core` instead.