Conditionally render config-driven properties per named tab
This commit moves "custom lead image" selection out of the main Document tab,
and into the Images tab. The property / tab assignment is driven by a new
property in the JSON - tab - which must be one of two enum values ("document",
or "images"). We can extend this later if we want to be able to surface custom
block fields on different tabs (e.g. a dedicated social media accounts tab).

Translations ignore the property, however, and surface all custom block
properties on one edit screen. This is because translations do not have a
tabbed editing interface. We should probably explore giving them such an
interface one day, but it does raise questions around whether translations
should have their own scoped area for image/attachment uploads, and if not, how
we explain to the publisher that any uploaded content is scoped to the document
and not the translation. That's a bit of a can of worms and the low-cost way
forward is to simply surface all the properties on the one translation edit
form. It's worth noting that 'legacy' content types have no means of specifying
a lead image for a translation at all, so the resulting experience is still an
improvement for the publisher, insofar as they can specify a different lead
image for their translation if they wish (something that would be useful for
HMRC, who often have to put out bilingual content).

Why this change?
It's always been a bit of a workaround to force publishers to have to
upload their image in one tab but 'assign' it to a lead image status in the
other. This was done for simplicity at the time, but won't scale very well to
when we migrate Topical Event pages to be config-driven; for those pages, we
expect to be able to designate a 'header' and 'logo' image too, and the designs
we have for that workflow very much expect us to be able to assign those images
on the Images tab itself.

Moving the existing "Custom lead image" workaround to the Images tab is a
valuable first step in that direction.