Build the payload for lead image when content is nil
The default lead image behaviour implies that we should send a payload including the default image url to publishing api, even when there is no custom image selection. Nonetheless, in the context of the block, the content only refers to the custom image, and that is what we save in whitehall. We don't really want to save the default image or the placeholder image reference. This would mean duplicating that data when we only need to send it to publishing api, but it also poses some technical difficulties, as we currently save the ImageData id for content. Attempting to save the default news image reference means we'd need to additionally store a FeaturedImageData reference, so we'd need a polymorphic relationship.

It's possible this is the implementation we are going to go for for future iterations. This is not the pre-existing behaviour though - i.e. WH only saved the custom image prior.

For now, we are allowing the payload call to run when the content is nil for lead image. Previous commit ensures there is always a placeholder to send in the payload.

This was not caught in testing (i.e. the image is not set to the default url in the content item), due to the page rendering as expected for news article, because this logic is currently duplicated on FE. Nonetheless, the logic is incomplete in FE so it will not work as expected for world news stories.