Allow users to see a document's current slug on summary page
Given increased support requests and confusion around when a document's slug changes, we are now letting users see the current 'full path' of a document when they visit the summary page. This is also in line with the new flow of allowing a document's slug to change once the title is updated.

JIRA: https://gov-uk.atlassian.net/browse/WHIT-1889