Add NewsStoryRecipe (inheriting from NewsArticleRecipe)
This defines the 'recipe' for converting a legacy news story into
a config-driven / StandardEdition news story.

Premature optimisiation perhaps, but I've gone straight for
defining most of the recipe logic in the NewsArticleRecipe and
then having NewsStoryRecipe inherit from it. This is because we
know that government responses and press releases are basically
identical to news stories, so will follow the same recipe, and
we don't want to have to duplicate lots of logic for each one.

World news stories are a little different, and we expect to be
able to define a WorldNewsStoryRecipe that mostly defers to
NewsArticleRecipe (via calls to `super`), but will need a little
bit of customisation too.