Convert public_updated_at to rfc3339 format [WHIT-2638]
This is what the legacy News Article presenter does. It doesn't
appear we deliberately chose _not_ to convert to this format in
the original implementation of the Standard Edition presenter,
and the version prior to this commit does appear to work.

However, keeping the two different arbitrarily is unhelpful for the
'standard edition migrator' service I'm building as part of this
ticket. When it comes to timestamps it is probably best to be
explicit about the format.

Jira: https://gov-uk.atlassian.net/browse/WHIT-2638