Only compare before-and-after payloads on the latest edition
We have to account for documents whose news article type ID _changes_ in
its document history, e.g.

```
Edition.find(545955).document.editions.pluck(:news_article_type_id)
=> [1, 1, 1, 1, 1, 2, 2]
```

In the example above, the document started out as a news story but was
later changed to a press release. By doing a strict payload comparison
on each edition, the migration of this document would fail:

- Choosing NewsStory recipe would cause it to fail after it had processed
  the first 5 editions
- Choosing PressRelease recipe would cause it to fail immediately (before
  it could get to the two press release editions)

In reality, we only care about the _latest_ edition and that its payload
is what we expect after we've migrated it. We only care about past editions
insofar as we want to be able to go back and see the superseded editions
etc and not worry about Whitehall crashing because they're of a now-deleted
document type. But we never send the payloads of old editions to Publishing
API, so we don't need to do a strict comparison check here.

Dropping that payload comparison check has the happy side-effect of making
it MUCH faster to migrate an individual document. In a local test run I
quickly found a document which had over 450 unique editions, which was slow
to migrate as each edition needed its associations expanded in order to
populate the Publishing API payload 'links' and so on. Now that their
migration operation has been simplified to just updating their row in the
DB, the operation is orders of magnitude faster.