Add StandardEditionMigratorWorker
This worker will take a document ID, find the document's editions
(including deleted ones), and migrate it to the StandardEdition
format, following a 'recipe' that will be defined in the coming
commits.

It also compares the payloads of each edition before and after
migration, raising an exception if they differ.